+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### [13.20.1](https://kaos.sh/ek/13.20.1)
 
+- **`[timeutil]`** Added method `Period.Stringf`
+- **`[timeutil]`** `Period.String` returns period range instead of duration
 - **`[log]`** Fixed bug with fields color for debug messages
 
 ### [13.20.0](https://kaos.sh/ek/13.20.0)